+++ src/wiremind_kubernetes/tests/e2e_tests/create_job_test.py
@@ -1,6 +1,8 @@
 import logging
 import time
 
+import kubernetes
+
 from .conftest import TEST_NAMESPACE
 
 logger = logging.getLogger(__name__)
@@ -8,7 +10,7 @@ logger = logging.getLogger(__name__)
 
 def test_create_job(concerned_dm, create_namespace):
     """
-    Test that default create job works as expected
+    Test that default create job and delete job work as expected
     """
     job_name = "my-test-job"
     concerned_dm.create_job(
@@ -26,6 +28,30 @@ def test_create_job(concerned_dm, create_namespace):
             logger.info("job not ready yet, waiting...")
             time.sleep(5)
     assert created_job.status.active == 1
+
+    concerned_dm.delete_job(job_name=job_name)
+    for _ in range(1, 10):
+        try:
+            concerned_dm.client_batchv1_api.read_namespaced_job(
+                concerned_dm.release_name + "-" + job_name, TEST_NAMESPACE
+            )
+        except kubernetes.client.rest.ApiException as e:
+            if e.status == 404:
+                break
+        else:
+            logger.info("job is not deleted yet, waiting...")
+            time.sleep(1)
+
+    for _ in range(1, 10):
+        pod_list = concerned_dm.client_corev1_api.list_namespaced_pod(
+            TEST_NAMESPACE, label_selector=f"job-name={concerned_dm.release_name}-{job_name}"
+        ).items
+        if not pod_list:
+            break
+        else:
+            logger.info("Related pod is not deleted yet, waiting...")
+            time.sleep(1)
+    assert not pod_list
 
 
 def test_create_job_argument(concerned_dm, create_namespace):